requirements.txt
click>=8.1.8
PyYAML>=6.0.2
PyGithub>=2.1.1
python-gitlab>=5.6.0